fix btn in show + authorization index

+++ app/helpers/check_show_link_to_helper.rb
@@ -1,14 +1,13 @@
 module CheckShowLinkToHelper
-  def check_show
-    if current_user && current_user.id == @list.user_id
-      link_to "delete", list_path(@list), method: :delete
+  def check_shotgun_btn
+    if current_user && @list.belongs_to_user?(current_user)
+      link_to "Delete my shotgun", list_shotgun_path(@list.id), method: :delete
     else
-      if current_user && @list.belongs_to_user?(current_user)
-        # Bouton delete my shotgunz à implanter
-        link_to "Delete my shotgun", list_shotgun_path(@list.id), method: :delete
-      else
-        link_to "Shotgun !", list_shotguns_path(@list.id), method: :post
-      end
+      link_to "Shotgun !", list_shotguns_path(@list.id), method: :post
     end
+  end
+
+  def check_delete_btn
+    link_to "delete", list_path(@list), method: :delete if current_user && current_user.id == @list.user_id
   end
 end
